+++ sdk/attestation/azure-security-attestation/samples/basic-operations/create_client.cpp
@@ -35,7 +35,7 @@ int main()
     clientOptions.TokenValidationOptions.TimeValidationSlack = 10s;
 
     // create client
-    AttestationClient const attestationClient(
+    AttestationClient attestationClient(
         AttestationClient::Create(GetEnvHelper::GetEnv("ATTESTATION_AAD_URL"), clientOptions));
 
     attestationClient.GetOpenIdMetadata();