+++ sdk/core/azure-core/CHANGELOG.md
@@ -1,16 +1,24 @@
 # Release History
 
-## 1.7.0-beta.1 (Unreleased)
+## 1.7.0-beta.2 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
 
-- Removed `noexcept` specification from `Azure::Core::Context::IsCancelled()`.
-
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.7.0-beta.1 (2022-06-02)
+
+### Features Added
+
+- Added prototypes and initial service support for Distributed Tracing.
+
+### Breaking Changes
+
+- Removed `noexcept` specification from `Azure::Core::Context::IsCancelled()`.
 
 ## 1.6.0 (2022-05-05)