+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.3.0-beta.3 (Unreleased)
+## 1.4.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,13 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.3.0 (2022-06-07)
+
+### Features Added
+
+- Added `ClientCertificateCredential`, and updated `EnvironmentCredential` to support client certificate authentication.
+- Added `ChainedTokenCredential`.
 
 ## 1.3.0-beta.2 (2022-05-10)